Fix --format CLI flag (#248)
* Fix --format CLI flag

Previously, the CLI used clap's PossibleValueParser, which only supports
parsing strings. However, hayagriva was expecting to parse Format
directly. This caused a panic in clap, when trying to convert the String
to Format.

With this patch, Format implements ValueEnum, so we can use
clap::value_parser to create an EnumValueParser to parse the Format.

* remove from_str from format

---------

Co-authored-by: PgBiel <9021226+PgBiel@users.noreply.github.com>