Stop `chapter` from breaking BibTeX-based entries
`chapter` in BibLaTeX represents "a chapter or section or any other unit
of a work" according to its manual [1][biblatex].  In the BibTeX manual,
it's said to be "a chapter (or section or whatever) number" [2][bibtex].

I understand this to mean that in Bib(La)TeX land, `chapter` is
something akin to `pages`, an indication to the curious reader where to
look to find the source.  Suffice to say, the current state where near
all other bibliographic information is thrown away to make way for an
`EntryType::Chapter` misunderstands the semantics of `chapter` (see
\#357)

I believe to achieve the same semantics as Bib(La)TeX, a new field would
need to be added to `Entry` to represent a chapter/section number.  I'll
leave that to someone else to figure out.  Until then, this change
prevents `chapter` from borking your bibliography entry.

Closes #357

[biblatex]: https://www.texlive.info/CTAN/macros/latex/contrib/biblatex/doc/biblatex.pdf
[bibtex]: https://www.nic.funet.fi/pub/TeX/CTAN/biblio/bibtex/base/btxdoc.pdf